Second parameter to Range.setStart() / setEnd() should be mandatory https://bugs.webkit.org/show_bug.cgi?id=160184
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

Rebaseline W3C test now that more checks are passing.

* web-platform-tests/dom/interfaces-expected.txt:

Source/WebCore:

Second parameter to Range.setStart() / setEnd() should be mandatory:
- https://dom.spec.whatwg.org/#interface-range

Also use "unsigned long" instead of "long" type for the parameter,
as per the specification.

Firefox and Chrome agree with the specification.

No new tests, rebaselined existing test.

* dom/Range.cpp:
(WebCore::Range::setStart):
(WebCore::Range::setEnd):
(WebCore::Range::checkNodeWOffset):
* dom/Range.h:
* dom/Range.idl:
* dom/RangeBoundaryPoint.h:
(WebCore::RangeBoundaryPoint::set):

LayoutTests:

Update tests to reflect behavior change.

* editing/deleting/delete-uneditable-style.html:
* fast/dom/non-numeric-values-numeric-parameters-expected.txt:
* fast/dom/script-tests/non-numeric-values-numeric-parameters.js:
* fast/regions/simplified-layout-no-regions.html:


Canonical link: https://commits.webkit.org/178342@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203713 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/dom/Range.cpp
@@ -160,7 +160,7 @@ static inline bool checkForDifferentRootContainer(const RangeBoundaryPoint& star
     return startRootContainer != endRootContainer || (Range::compareBoundaryPoints(start, end, ASSERT_NO_EXCEPTION) > 0);
 }
 
-void Range::setStart(Ref<Node>&& refNode, int offset, ExceptionCode& ec)
+void Range::setStart(Ref<Node>&& refNode, unsigned offset, ExceptionCode& ec)
 {
     bool didMoveDocument = false;
     if (&refNode->document() != &ownerDocument()) {
@@ -179,7 +179,7 @@ void Range::setStart(Ref<Node>&& refNode, int offset, ExceptionCode& ec)
         collapse(true);
 }
 
-void Range::setEnd(Ref<Node>&& refNode, int offset, ExceptionCode& ec)
+void Range::setEnd(Ref<Node>&& refNode, unsigned offset, ExceptionCode& ec)
 {
     bool didMoveDocument = false;
     if (&refNode->document() != &ownerDocument()) {
@@ -921,7 +921,7 @@ void Range::detach()
     // This is now a no-op as per the DOM specification.
 }
 
-Node* Range::checkNodeWOffset(Node& node, int offset, ExceptionCode& ec) const
+Node* Range::checkNodeWOffset(Node& node, unsigned offset, ExceptionCode& ec) const
 {
     switch (node.nodeType()) {
         case Node::DOCUMENT_TYPE_NODE:
@@ -931,7 +931,7 @@ Node* Range::checkNodeWOffset(Node& node, int offset, ExceptionCode& ec) const
         case Node::COMMENT_NODE:
         case Node::TEXT_NODE:
         case Node::PROCESSING_INSTRUCTION_NODE:
-            if (static_cast<unsigned>(offset) > downcast<CharacterData>(node).length())
+            if (offset > downcast<CharacterData>(node).length())
                 ec = INDEX_SIZE_ERR;
             return nullptr;
         case Node::ATTRIBUTE_NODE:
